Move shareable libs into their own packages.
