+++ go.mod
@@ -4,6 +4,7 @@ go 1.13
 
 require (
 	github.com/Knetic/govaluate v3.0.1-0.20170526162915-ed428fc1d628+incompatible
+	github.com/franela/goblin v0.0.0-20181003173013-ead4ad1d2727 // indirect
 	github.com/franela/goreq v0.0.0-20171204163338-bcd34c9993f8
 	github.com/golang/protobuf v1.3.2
 	github.com/gorilla/handlers v1.4.2
@@ -12,6 +13,7 @@ require (
 	github.com/hashicorp/go-cleanhttp v0.5.1
 	github.com/hashicorp/serf v0.8.5
 	github.com/jtolds/gls v4.20.0+incompatible
+	github.com/onsi/gomega v1.7.1 // indirect
 	github.com/op/go-logging v0.0.0-20160315200505-970db520ece7
 	github.com/smartystreets/assertions v1.0.1 // indirect
 	github.com/smartystreets/goconvey v1.6.4
