+++ pkg/columnizer/columnize.go
@@ -1,4 +1,4 @@
-package main
+package columnizer
 
 import (
 	"fmt"
