+++ pkg/columnizer/columnize_test.go
@@ -1,4 +1,4 @@
-package main
+package columnizer
 
 import (
 	"bytes"
